wayland/xcb: fix client not returning VkPhysical device
The issue was due to enter &=. If INCLUDE_KMS isn't
included and enter=0 this will leave the value of enter
at zero.

If INCLUDE_KMS is included commit ensures even if we have a
valid DRI device node fd. That the user choosen device type
is the same as the DRI device node.

Signed-off-by: Vincent Davis Jr <vince@underview.tech>