vulkan: struct uvr_vk_descriptor_set_create_info member rename
Renamed member @descriptorPoolSetsInfo to @descriptorPoolInfos
and added comments so that it hopefully makes more sense in the
future. Also added @descriptorSetLayoutCount member. This is
used to determine the amount of descriptor sets a descriptor pool
can allocate. It's also used to determine the array size of
@descriptorSetLayouts. So, when descriptors are assigned to a
set vulkan knows which descriptors belong to which descriptor set.

Signed-off-by: Vincent Davis Jr <vince@underview.tech>