examples: update texture image format
Updates format to VK_FORMAT_R8G8B8A8_SRGB. This format
appears to get the best results in terms of how similar
to the original image it looks. Assumption is that
stb_image creates pixel buffers in RGBA format.
Specifying in the VkImage associated with texture
that it's format should be RGBA yeilds the correct
results.

Signed-off-by: Vincent Davis Jr <vince@underview.tech>