examples[model-load]: fragment shader colors all GLTF baseColorTexture's
Final image appears to be close to the image in Sascha Willems
Vulkan API. Needed to update uniform shader variable
with the index of the texture bound to the descriptor.

It's implemented this way because I only have one
descriptor set and many descriptors bound to that
set (7 descriptors total, 1 descriptor set). The first
two descriptors being uniform descriptors and the last
five being combinded image sampler descriptors.

These five descriptors are bound to our GLTF file
material.baseColorTextureIndex. Which of course
is associated with an image texture. Needed
to create sampler2D array in the fragment shader
then using the index fill in our textures color.

Signed-off-by: Vincent Davis Jr <vince@underview.tech>