kms: fix gbm buffer create segfault if libseat not used
Issue was that the file descriptor associated with
the GPU was created with the O_RDONLY flag. It appears
that if the file descriptor is READONLY the driver
can't allocate gbm buffers. The application actually
sefaults.

Create the file descriptor associated with the GPU
with the O_RDWR so the driver can create GPU buffers.

Signed-off-by: Vincent Davis Jr <vince@underview.tech>